moved code into Ingest as an option
