--- cosmos/ingestion/ingest/enrich/tests/test_enrich.py
@@ -1,36 +1,0 @@
-from unittest import TestCase
-
-
-class TestEnrich(TestCase):
-    def test_get_file_structure(self):
-        self.fail()
-
-    def test_get_score_histogram(self):
-        self.fail()
-
-    def test_show_table_rows(self):
-        self.fail()
-
-    def test_semantic_enrichment(self):
-        self.fail()
-
-    def test_set_table_ids(self):
-        self.fail()
-
-    def test_apply_table_labels(self):
-        self.fail()
-
-    def test_create_semantic_contexts(self):
-        self.fail()
-
-    def test_get_semantic_context(self):
-        self.fail()
-
-    def test_get_labels_and_spans_from_content(self):
-        self.fail()
-
-    def test_get_labels_and_sentences_with_regex(self):
-        self.fail()
-
-    def test_export_data(self):
-        self.fail()
